# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "API documentation for Laravel application",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/company-store": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Company Store",
                "operationId": "dbb67f351460391427aaa69ac1357f2b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employerFullName": {
                                        "description": "Full Name of Employer",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "employerEmail": {
                                        "description": "Email of Employer",
                                        "type": "string",
                                        "example": "john@gmail.com"
                                    },
                                    "employerPassword": {
                                        "description": "Password of Employer",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "employerConfirmPassword": {
                                        "description": "Confirm Password of Employer",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "companyName": {
                                        "description": "Full name of Company",
                                        "type": "string",
                                        "example": "Tech Company"
                                    },
                                    "companyEmail": {
                                        "description": "Email of Company",
                                        "type": "string",
                                        "example": "company@email.com"
                                    },
                                    "companyAddress": {
                                        "description": "Address of Company",
                                        "type": "string",
                                        "example": "Tech Company"
                                    },
                                    "companyPhone": {
                                        "description": "Phone of Company",
                                        "type": "string",
                                        "example": "Tech Company"
                                    },
                                    "companyDescription": {
                                        "description": "Description of Company",
                                        "type": "string",
                                        "example": "Tech Company"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "company_name": {
                                            "description": "Name of Company",
                                            "type": "string",
                                            "example": "tech Company"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Unsucessful"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/super-admin/company-activation": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Activate Company",
                "operationId": "1db720af930e03b02fd343d90312e6da",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "companyId": {
                                        "description": "Id of Company",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "activationStatus": {
                                        "description": "Activation Status of Company",
                                        "type": "boolean",
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "employerId": {
                                            "description": "Id of Employer",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "employerEmail": {
                                            "description": "Email of Employer",
                                            "type": "string",
                                            "example": "company@gmail.com"
                                        },
                                        "companyId": {
                                            "description": "Id of Company",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "companyStatus": {
                                            "description": "Status of Company",
                                            "type": "boolean",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Unsucessful"
                    },
                    "400": {
                        "description": "Repeated Action"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}